refactor: use the same cache instance for Tests and Runs
It's backed by the same storage, no need to create separate instances per action kind.
