+++ core/src/plugins/kubernetes/container/run.ts
@@ -9,12 +9,11 @@
 import type { ContainerRunAction } from "../../container/moduleConfig.js"
 import { runAndCopy } from "../run.js"
 import type { KubernetesPluginContext } from "../config.js"
-import { getRunResultCache } from "../run-results.js"
 import { makePodName } from "../util.js"
 import { getNamespaceStatus } from "../namespace.js"
 import type { RunActionHandler } from "../../../plugin/action-types.js"
 import { getDeployedImageId } from "./util.js"
-import { composeCacheableResult, toActionStatus } from "../results-cache.js"
+import { composeCacheableResult, getResultCache, toActionStatus } from "../results-cache.js"
 
 export const k8sContainerRun: RunActionHandler<"run", ContainerRunAction> = async (params) => {
   const { ctx, log, action } = params
@@ -46,7 +45,7 @@ export const k8sContainerRun: RunActionHandler<"run", ContainerRunAction> = asyn
   const detail = composeCacheableResult({ result, namespaceStatus })
 
   if (action.getSpec("cacheResult")) {
-    const runResultCache = getRunResultCache(ctx.gardenDirPath)
+    const runResultCache = getResultCache(ctx.gardenDirPath)
     await runResultCache.store({
       ctx,
       log,
