+++ core/src/plugins/kubernetes/container/test.ts
@@ -7,14 +7,13 @@
  */
 
 import type { ContainerTestAction } from "../../container/moduleConfig.js"
-import { getTestResultCache } from "../test-results.js"
 import { runAndCopy } from "../run.js"
 import { makePodName } from "../util.js"
 import { getNamespaceStatus } from "../namespace.js"
 import type { KubernetesPluginContext } from "../config.js"
 import type { TestActionHandler } from "../../../plugin/action-types.js"
 import { getDeployedImageId } from "./util.js"
-import { composeCacheableResult, toActionStatus } from "../results-cache.js"
+import { composeCacheableResult, getResultCache, toActionStatus } from "../results-cache.js"
 
 export const k8sContainerTest: TestActionHandler<"run", ContainerTestAction> = async (params) => {
   const { ctx, log, action } = params
@@ -46,7 +45,7 @@ export const k8sContainerTest: TestActionHandler<"run", ContainerTestAction> = a
   const detail = composeCacheableResult({ result, namespaceStatus })
 
   if (action.getSpec("cacheResult")) {
-    const testResultCache = getTestResultCache(ctx.gardenDirPath)
+    const testResultCache = getResultCache(ctx.gardenDirPath)
     await testResultCache.store({
       ctx,
       log,
