+++ core/test/integ/src/plugins/kubernetes/container/run.ts
@@ -16,11 +16,12 @@ import fsExtra from "fs-extra"
 const { emptyDir, pathExists } = fsExtra
 import { join } from "path"
 import { getContainerTestGarden } from "./container.js"
-import { getRunResultCache } from "../../../../../../src/plugins/kubernetes/run-results.js"
 import type { KubernetesProvider } from "../../../../../../src/plugins/kubernetes/config.js"
 import type { ContainerRunAction } from "../../../../../../src/plugins/container/config.js"
 import { createActionLog } from "../../../../../../src/logger/log-entry.js"
 import { waitForOutputFlush } from "../../../../../../src/process.js"
+
+import { getResultCache } from "../../../../../../src/plugins/kubernetes/results-cache.js"
 
 describe("runContainerTask", () => {
   let garden: TestGarden
@@ -58,7 +59,7 @@ describe("runContainerTask", () => {
     garden.events.eventLog = []
 
     const ctx = await garden.getPluginContext({ provider, templateContext: undefined, events: undefined })
-    const runResultCache = getRunResultCache(ctx.gardenDirPath)
+    const runResultCache = getResultCache(ctx.gardenDirPath)
     await runResultCache.clear({ ctx, log: garden.log, action })
 
     const results = await garden.processTasks({ tasks: [testTask], throwOnError: true })
@@ -107,7 +108,7 @@ describe("runContainerTask", () => {
     })
 
     const ctx = await garden.getPluginContext({ provider, templateContext: undefined, events: undefined })
-    const runResultCache = getRunResultCache(ctx.gardenDirPath)
+    const runResultCache = getResultCache(ctx.gardenDirPath)
     await runResultCache.clear({ ctx, log: garden.log, action })
 
     await garden.processTasks({ tasks: [testTask], throwOnError: true })
@@ -144,7 +145,7 @@ describe("runContainerTask", () => {
     })
 
     const ctx = await garden.getPluginContext({ provider, templateContext: undefined, events: undefined })
-    const runResultCache = getRunResultCache(ctx.gardenDirPath)
+    const runResultCache = getResultCache(ctx.gardenDirPath)
     await runResultCache.clear({ ctx, log: garden.log, action })
 
     await expectError(
