+++ core/test/integ/src/plugins/kubernetes/kubernetes-type/kubernetes-exec-run.ts
@@ -13,7 +13,8 @@ import { findNamespaceStatusEvent } from "../../../../../helpers.js"
 import type { ConfigGraph } from "../../../../../../src/graph/config-graph.js"
 import { getKubernetesTestGarden } from "./common.js"
 import { RunTask } from "../../../../../../src/tasks/run.js"
-import { getRunResultCache } from "../../../../../../src/plugins/kubernetes/run-results.js"
+
+import { getResultCache } from "../../../../../../src/plugins/kubernetes/results-cache.js"
 
 describe("kubernetes-type exec Run", () => {
   let garden: TestGarden
@@ -42,7 +43,7 @@ describe("kubernetes-type exec Run", () => {
     // Clear any existing Run result
     const provider = await garden.resolveProvider({ log: garden.log, name: "local-kubernetes" })
     const ctx = await garden.getPluginContext({ provider, templateContext: undefined, events: undefined })
-    const runResultCache = getRunResultCache(ctx.gardenDirPath)
+    const runResultCache = getResultCache(ctx.gardenDirPath)
     await runResultCache.clear({ ctx, log: garden.log, action })
 
     garden.events.eventLog = []
@@ -72,7 +73,7 @@ describe("kubernetes-type exec Run", () => {
     // Clear any existing Run result
     const provider = await garden.resolveProvider({ log: garden.log, name: "local-kubernetes" })
     const ctx = await garden.getPluginContext({ provider, templateContext: undefined, events: undefined })
-    const runResultCache = getRunResultCache(ctx.gardenDirPath)
+    const runResultCache = getResultCache(ctx.gardenDirPath)
     await runResultCache.clear({ ctx, log: garden.log, action })
 
     garden.events.eventLog = []
