+++ core/test/integ/src/plugins/kubernetes/kubernetes-type/kubernetes-pod-run.ts
@@ -16,9 +16,10 @@ import { RunTask } from "../../../../../../src/tasks/run.js"
 import fsExtra from "fs-extra"
 const { emptyDir, pathExists } = fsExtra
 import { join } from "path"
-import { getRunResultCache } from "../../../../../../src/plugins/kubernetes/run-results.js"
 import type { KubernetesPodRunAction } from "../../../../../../src/plugins/kubernetes/kubernetes-type/kubernetes-pod.js"
 import { createActionLog } from "../../../../../../src/logger/log-entry.js"
+
+import { getResultCache } from "../../../../../../src/plugins/kubernetes/results-cache.js"
 
 describe("kubernetes-type pod Run", () => {
   let garden: TestGarden
@@ -47,7 +48,7 @@ describe("kubernetes-type pod Run", () => {
     // Clear any existing Run result
     const provider = await garden.resolveProvider({ log: garden.log, name: "local-kubernetes" })
     const ctx = await garden.getPluginContext({ provider, templateContext: undefined, events: undefined })
-    const runResultCache = getRunResultCache(ctx.gardenDirPath)
+    const runResultCache = getResultCache(ctx.gardenDirPath)
     await runResultCache.clear({ ctx, log: garden.log, action })
 
     garden.events.eventLog = []
@@ -89,7 +90,7 @@ describe("kubernetes-type pod Run", () => {
     // Clear any existing Run result
     const provider = await garden.resolveProvider({ log: garden.log, name: "local-kubernetes" })
     const ctx = await garden.getPluginContext({ provider, templateContext: undefined, events: undefined })
-    const runResultCache = getRunResultCache(ctx.gardenDirPath)
+    const runResultCache = getResultCache(ctx.gardenDirPath)
     await runResultCache.clear({ ctx, log: garden.log, action })
 
     await garden.processTasks({ tasks: [runTask], throwOnError: true })
